avance ventas

+++ PrjFarmacia/src/org/farmacia/dao/ClienteDao.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import org.farmacia.bean.Cliente;
 import org.farmacia.dao.base.BaseDao;
 
-public interface ClienteDao extends BaseDao<Cliente, Integer>{
+public interface ClienteDao extends BaseDao<Cliente, Long>{
 	public void crearCliente(Cliente usuario);
 	public List<Cliente> listarClientesPorParametroBusqueda(Map<String, Object> parametros);
 }
